requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.4.1
bcrypt==3.2.2
cffi==1.15.0
click==8.0.4
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
cryptography==37.0.2
dataclasses==0.8
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
jose==1.0.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing==3.7.4.3
typing_extensions==4.1.1
uvicorn==0.16.0
zipp==3.6.0
